relocated the optimization code into the specific scripts with updated documentation. next is to put udapte the init scripts outside of the optimization module and add a read me
